requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
lightgbm>=3.3.0
xgboost>=1.7.0
catboost>=1.1.0
shap>=0.41.0
optuna>=3.0.0
pyyaml>=6.0
matplotlib>=3.4.0
seaborn>=0.11.0
tqdm>=4.62.0
mlflow>=2.0.0
pytest>=7.0.0
pytest-cov>=3.0.0
joblib>=1.1.0
scipy>=1.7.0
